resolved comments

+++ src/main/java/com/rakovets/course/java/core/practice/oop_classes_and_objects/Utils.java
@@ -1,8 +1,7 @@
 package com.rakovets.course.java.core.practice.oop_classes_and_objects;
 
 public class Utils {
-    public static int getSquareOfNumber(int number) {
-        return (int) Math.pow(number, 2);
+    public static double getSquareOfNumber(double number) {
+        return Math.pow(number, 2);
     }
-
 }
